Reorganize Comment files
